+++ tockn/view/src/store/modules/gists/actions.js
@@ -1,4 +1,5 @@
 import axios from 'axios'
+import {request} from '../../utils'
 const API_ENDPOINT = process.env.API_ENDPOINT
 
 export default {
@@ -37,15 +38,4 @@ export default {
         commit('checkStarred', response)
       })
   }
-}
-
-var request = function (method, url, accessToken) {
-  let reqObj = {
-    method: method,
-    url: url
-  }
-  if (accessToken !== '') {
-    reqObj.headers = {'Authorization': `bearer ${accessToken}`}
-  }
-  return reqObj
 }
